added a11y and tests to the mega menu dropdown
